refactor: organise all routes into separate directory
