+++ server/routes/__init__.py
@@ -1,0 +1,3 @@
+from routes.api_routes import api_routes
+from routes.core_routes import core_routes
+from routes.dev_routes import dev_routes
